NOTE -- updating DXY data source
BlankerL has generously provided API hosting, but the bandwidth he now recieves is overwhelming the self-hosted site (https://github.com/BlankerL/DXY-COVID-19-Crawler/issues/53).

Until I can clone and implement the DB and webscraper, we will be pointing to the updated JSON files as an upstream source. These are directly output by BlankerL as well, so that should be suitable.